+++ apps/docs/components/docs/components/codeblock.tsx
@@ -1,10 +1,9 @@
 import React, {forwardRef, useEffect} from "react";
 import {clsx, dataAttr, getUniqueID} from "@nextui-org/shared-utils";
 import BaseHighlight, {Language, PrismTheme, defaultProps} from "prism-react-renderer";
-import {debounce} from "lodash";
+import {debounce, omit} from "lodash";
 
 import defaultTheme from "@/libs/prism-theme";
-import {trackEvent} from "@/utils/va";
 
 interface CodeblockProps {
   language: Language;
@@ -20,6 +19,17 @@ interface CodeblockProps {
 
 type HighlightStyle = "inserted" | "deleted" | undefined;
 
+const nextuiCliCommand = [
+  /^init$/,
+  /^add$/,
+  /^upgrade$/,
+  /^remove$/,
+  /^list$/,
+  /^env$/,
+  /^doctor$/,
+];
+
+const highlightStyleToken = ["bun", /nextui\s\w+(?=\s?)/, /^nextui$/, "Usage", ...nextuiCliCommand];
 const RE = /{([\d,-]+)}/;
 
 const calculateLinesToHighlight = (meta?: string) => {
@@ -112,12 +122,6 @@ const Codeblock = forwardRef<HTMLPreElement, CodeblockProps>(
           return;
 
         lastSelectionText.current = selectionText;
-
-        trackEvent("Codeblock - Selection", {
-          action: "selectText",
-          category: "docs",
-          data: selectionText,
-        });
       };
 
       const debouncedHandleSelectionChange = debounce(handleSelectionChange, 1000);
@@ -152,8 +156,8 @@ const Codeblock = forwardRef<HTMLPreElement, CodeblockProps>(
 
                 return (
                   <div
+                    {...omit(lineProps, ["key"])}
                     key={`${i}-${getUniqueID("line-wrapper")}`}
-                    {...lineProps}
                     className={clsx(
                       lineProps.className,
                       removeIndent ? "pr-4" : "px-4",
@@ -172,13 +176,28 @@ const Codeblock = forwardRef<HTMLPreElement, CodeblockProps>(
                     {showLines && (
                       <span className="select-none text-xs mr-6 opacity-30">{i + 1}</span>
                     )}
-                    {line.map((token, key) => (
-                      <span
-                        key={`${key}-${getUniqueID("line")}`}
-                        {...getTokenProps({token, key})}
-                        className={className}
-                      />
-                    ))}
+                    {line.map((token, key) => {
+                      // Bun has no color style by default in the code block, so hack add in here
+                      const props = getTokenProps({token, key}) || {};
+
+                      return (
+                        <span
+                          {...omit(props, ["key"])}
+                          key={`${key}-${getUniqueID("line")}`}
+                          className={className}
+                          style={{
+                            ...props.style,
+                            ...(highlightStyleToken.some((t) => {
+                              const regex = t instanceof RegExp ? t : new RegExp(t);
+
+                              return regex.test(token.content.trim());
+                            })
+                              ? {color: "rgb(var(--code-function))"}
+                              : {}),
+                          }}
+                        />
+                      );
+                    })}
                   </div>
                 );
               })}