+++ packages/components/dropdown/__tests__/dropdown.test.tsx
@@ -2,10 +2,21 @@ import * as React from "react";
 import {act, render} from "@testing-library/react";
 import {Button} from "@nextui-org/button";
 import userEvent from "@testing-library/user-event";
+import {User} from "@nextui-org/user";
+import {Image} from "@nextui-org/image";
+import {Avatar} from "@nextui-org/avatar";
 
 import {Dropdown, DropdownTrigger, DropdownMenu, DropdownItem, DropdownSection} from "../src";
 
+// e.g. console.error Warning: Function components cannot be given refs.
+// Attempts to access this ref will fail. Did you mean to use React.forwardRef()?
+const spy = jest.spyOn(console, "error").mockImplementation(() => {});
+
 describe("Dropdown", () => {
+  afterEach(() => {
+    jest.clearAllMocks();
+  });
+
   it("should render correctly (static)", () => {
     const wrapper = render(
       <Dropdown>
@@ -100,6 +111,7 @@ describe("Dropdown", () => {
               items={section.children}
               title={section.title}
             >
+              {/* @ts-ignore */}
               {(item: any) => <DropdownItem key={item.key}>{item.name}</DropdownItem>}
             </DropdownSection>
           )}
@@ -108,6 +120,38 @@ describe("Dropdown", () => {
     );
 
     expect(() => wrapper.unmount()).not.toThrow();
+  });
+
+  it("should not throw any error when clicking button", async () => {
+    const wrapper = render(
+      <Dropdown>
+        <DropdownTrigger>
+          <Button data-testid="trigger-test">Trigger</Button>
+        </DropdownTrigger>
+        <DropdownMenu aria-label="Actions" onAction={alert}>
+          <DropdownItem key="new">New file</DropdownItem>
+          <DropdownItem key="copy">Copy link</DropdownItem>
+          <DropdownItem key="edit">Edit file</DropdownItem>
+          <DropdownItem key="delete" color="danger">
+            Delete file
+          </DropdownItem>
+        </DropdownMenu>
+      </Dropdown>,
+    );
+
+    let triggerButton = wrapper.getByTestId("trigger-test");
+
+    expect(triggerButton).toBeTruthy();
+
+    await act(async () => {
+      await userEvent.click(triggerButton);
+    });
+
+    expect(spy).toBeCalledTimes(0);
+
+    let menu = wrapper.queryByRole("menu");
+
+    expect(menu).toBeTruthy();
   });
 
   it("should work with single selection (controlled)", async () => {
@@ -353,6 +397,36 @@ describe("Dropdown", () => {
     expect(menu).toBeFalsy();
   });
 
+  it("should not open on disabled dropdown", () => {
+    const wrapper = render(
+      <Dropdown isDisabled>
+        <DropdownTrigger>
+          <Button data-testid="trigger-test">Trigger</Button>
+        </DropdownTrigger>
+        <DropdownMenu aria-label="Actions">
+          <DropdownItem key="new">New file</DropdownItem>
+          <DropdownItem key="copy">Copy link</DropdownItem>
+          <DropdownItem key="edit">Edit file</DropdownItem>
+          <DropdownItem key="delete" color="danger">
+            Delete file
+          </DropdownItem>
+        </DropdownMenu>
+      </Dropdown>,
+    );
+
+    let triggerButton = wrapper.getByTestId("trigger-test");
+
+    expect(triggerButton).toBeTruthy();
+
+    act(() => {
+      triggerButton.click();
+    });
+
+    let menu = wrapper.queryByRole("menu");
+
+    expect(menu).toBeFalsy();
+  });
+
   it("should not select on disabled item", () => {
     const onSelectionChange = jest.fn();
     const wrapper = render(
@@ -385,5 +459,82 @@ describe("Dropdown", () => {
     });
 
     expect(onSelectionChange).toBeCalledTimes(0);
+  });
+
+  it("should render without error (custom trigger with isDisabled)", async () => {
+    const spy = jest.spyOn(console, "error").mockImplementation(() => {});
+
+    const renderDropdownMenu = () => {
+      return (
+        <DropdownMenu aria-label="Actions" onAction={alert}>
+          <DropdownItem key="new">New file</DropdownItem>
+          <DropdownItem key="copy">Copy link</DropdownItem>
+          <DropdownItem key="edit">Edit file</DropdownItem>
+          <DropdownItem key="delete" color="danger">
+            Delete file
+          </DropdownItem>
+        </DropdownMenu>
+      );
+    };
+
+    // Non Next UI Element in DropdownTrigger
+    render(
+      <Dropdown isDisabled>
+        <DropdownTrigger>
+          <div>Trigger</div>
+        </DropdownTrigger>
+        {renderDropdownMenu()}
+      </Dropdown>,
+    );
+
+    expect(spy).toBeCalledTimes(0);
+
+    spy.mockRestore();
+
+    // Next UI Element in DropdownTrigger
+    render(
+      <Dropdown isDisabled>
+        <DropdownTrigger>
+          <User as="button" description="@tonyreichert" name="Tony Reichert" />
+        </DropdownTrigger>
+        {renderDropdownMenu()}
+      </Dropdown>,
+    );
+
+    expect(spy).toBeCalledTimes(0);
+
+    spy.mockRestore();
+
+    // NextUI Element that supports isDisabled prop in DropdownTrigger
+    render(
+      <Dropdown isDisabled>
+        <DropdownTrigger>
+          <Avatar isDisabled name="Marcus" />
+        </DropdownTrigger>
+        {renderDropdownMenu()}
+      </Dropdown>,
+    );
+
+    expect(spy).toBeCalledTimes(0);
+
+    spy.mockRestore();
+
+    // NextUI Element that doesn't support isDisabled prop in DropdownTrigger
+    render(
+      <Dropdown isDisabled>
+        <DropdownTrigger>
+          <Image
+            alt="NextUI hero Image"
+            src="https://nextui-docs-v2.vercel.app/images/hero-card-complete.jpeg"
+            width={300}
+          />
+        </DropdownTrigger>
+        {renderDropdownMenu()}
+      </Dropdown>,
+    );
+
+    expect(spy).toBeCalledTimes(0);
+
+    spy.mockRestore();
   });
 });